+++ data/src/main/java/com/yapp/growth/data/di/DataSourceModule.kt
@@ -2,7 +2,7 @@ package com.yapp.growth.data.di
 
 import com.yapp.growth.data.source.ConfirmPlanDataSource
 import com.yapp.growth.data.source.HomeDataSource
-import com.yapp.growth.data.source.HomeDataSourceImpl
+import com.yapp.growth.data.internal.source.HomeDataSourceImpl
 import com.yapp.growth.data.internal.source.ConfirmPlanDataSourceImpl
 import com.yapp.growth.data.internal.source.CreateTimeTableDataSourceImpl
 import com.yapp.growth.data.source.CreateTimeTableDataSource
