+++ data/src/main/java/com/yapp/growth/data/internal/response/FixedPlanResponseImpl.kt
@@ -1,4 +1,6 @@
-package com.yapp.growth.data.response
+package com.yapp.growth.data.internal.response
+
+import com.yapp.growth.data.response.FixedPlanResponse
 
 data class FixedPlanResponseImpl(
     override val id: Long,
@@ -10,13 +12,3 @@ data class FixedPlanResponseImpl(
     override val members: List<UserResponseImpl>,
     override val placeName: String
 ): FixedPlanResponse
-
-data class UserResponseImpl (
-    override val id: Long,
-    override val userName: String
-): UserResponse
-
-data class CategoryResponseImpl (
-    override val id: Long,
-    override val keyword: String
-): CategoryResponse
